templates: Add hvm ndvm templates
Compared to the pv version:

set hvm to true and enable stubdoms

boot=c to boot off the hard disk

qemu-dm-path needs to be set if running without stubdom

v4v-firewall rule for stubdom to talk to the qmp-helper

kernel-extract and command line removed  they are unneeded for hvm and
cause libxl to throw an error

display=nogfx to disable graphics

OXT-1327

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>

---

Memory gets decrease by 16MB, so we may want to bump.  Stable-6 seemed
to need a memory bump, but stable-8 seems okay with this lower value.