rpc-proxy: Lower EOF verbosity level
EOF is an unexceptional exception for the normal operation of rpc-proxy.
It occurs when a channel is closed normally.  As such, it only makes
sense to print at the debug level, but other exceptions truly are
exceptional and warn level makes sense.  Special case EOF to only print
at the lower level.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>